refactor: внесение изменений из main
