+++ packages/client/package.json
@@ -20,7 +20,6 @@
     "@tailwindcss/vite": "^4.1.8",
     "axios": "^1.9.0",
     "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "dotenv": "^16.0.2",
     "eslint-config-prettier": "^8.5.0",
@@ -45,13 +44,17 @@
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
     "@vitejs/plugin-react": "^2.0.1",
+    "clsx": "^2.1.1",
     "eslint": "^8.23.0",
     "jest": "^28",
     "jest-environment-jsdom": "^29.0.1",
     "lefthook": "^1.3.9",
     "prettier": "^2.7.1",
+    "radix-ui": "^1.4.2",
     "react-router-dom": "^7.6.1",
     "react-routes": "^0.2.6",
+    "sass-embedded": "^1.89.2",
+    "shadcn": "^2.6.4",
     "ts-jest": "^28.0.8",
     "tw-animate-css": "^1.3.3",
     "typescript": "^4.8.2",
