+++ packages/client/src/pages/Forum/Forum.module.scss
@@ -1,4 +1,4 @@
-@use '../../index.css';
+@use '../../app/index.css';
 
 .forum {
   width: 100vw;
