+++ packages/client/src/shared/ui/index.ts
@@ -1,3 +1,4 @@
+export * from './Avatar'
 export * from './Button'
 export * from './Card'
 export * from './Command'
